Bring Bazel builds into compliance with upcoming breaking changes
This change makes our Bazel build compatible with all the migration-ready
`--incompatible` flags listed
[here](https://github.com/bazelbuild/bazel/releases/tag/7.0.0) for Bazel 7.0,
with a couple exceptions listed below. This was mostly just a matter of
updating some of our dependencies to more recent versions.

The two flags we cannot support yet are these:
 - `--incompatible_disallow_struct_provider_syntax`: Bazel itself appears not
   to support this one yet, see
   [here](https://github.com/bazelbuild/bazel/issues/19467).
 - `--incompatible_stop_exporting_language_modules`: waiting on support in
   `rules_jvm_external`, see
   [here](https://github.com/bazelbuild/rules_jvm_external/pull/1027).

I also deleted the broken `//:protobuf_nowkt` alias target, which was an
oversight on my part from a previous change.

PiperOrigin-RevId: 597076542