#rust #protobuf Implement custom Debug for C++ kernel
This change implements a custom Debug for messages, views and muts in the C++ kernel. Debug defers to proto2::Utf8Format.

It implements this only for the C++ kernel. We will need to pull in additional dependencies beyond minitables to implement it for UPB as well. This will be done at a later point.

PiperOrigin-RevId: 603028937