Use is_a? instead of kind_of?
If it walks like a duck, quacks like a duck - it might be a duck.

Rails ActiveSupport::TimeWithZone is a Time-alike and can de serialized
correctly, it works when you do:

    Google::Protobuf::Timestamp.new(Time.current)

But as the check in `convert.rb` is `kind_of?` it prevents assigning
`Time.current` to a Timestamp field of a message directly without
type-casting it into Ruby native Time.

Signed-off-by: Alexey Zapparov <alexey@zapparov.com>