Add a maximum block size for arena growth.
This should significantly reduce the size of large arenas. Previously, a large arena would nearly double in size if the most recent block filled up.  This could end up wasting large amounts of memory.  After this CL, we will waste at most the max block size, which defaults to 32k.

This more or less matches the behavior of the C++ arena.

PiperOrigin-RevId: 647802280