Enable internal hasbits for scalar proto3 implicit presence fields.
__N.B. This change is not intended to affect any observable behaviour.__

There are quite a few terminologies in the title:

* scalar: a field that is not repeated, not oneof, not extension, not lazy,
	  just a field with a simple primitive type (number or boolean), or
	  string/bytes.
* proto3: describes behaviour consistent to the "proto3" syntax.
          This is equivalent to `edition = "2023"` with
	  `option features.field_presence = IMPLICIT;`.
* implicit presence: describes behaviour consistent with "non-optional" fields
                     in proto3. This is described in more detail in
                     https://protobuf.dev/programming-guides/field_presence/#presence-in-proto3-apis

This change enables C++ proto3 objects to generate hasbits for proto3 scalar
fields. This code change might make certain codepaths negligibly more
efficient, but large improvement or regression is unlikely. A larger
performance impact is expected from generating hasbits for repeated fields --
this change will pave the way for future work there.

Hasbits in C++ will have slightly different semantics for implicit presence
fields. In the past, all hasbits are true field presence indicators. If the
hasbit is set, the field is guaranteed to be present; if the hasbit is unset,
the field is guaranteed to be missing.

This change introduces a new hasbit mode that I will call "hint hasbits". For
implicit presence fields, it may be possible to mutate the field and have it
end up as a zero field, especially with `mutable_foo` APIs. To handle those
cases correctly, we must do an additional check for field emptiness before
serializing the field onto the wire.

The wire parsing codepath is not affected.

This change affects C++ protobuf generated objects only. Implementation in
other languages are not affected.

PiperOrigin-RevId: 654074096