[ObjC] Collect unknown fields as data
Default to collecting unknown fields as data and then create the
`GPBUnknownFieldSet` on demand when requested.

This should reduce the as the extra objects are created until requested
and clears the way to eventually deprecated `GPBUnknownFieldSet` in the
future.

This also fixes the failing conformance test related to keeping the ordering.

PiperOrigin-RevId: 655929316