Formatting beautifications.
Put the emission of braces and indentation into MayEmitIfNonDefaultCheck, this
allows for a slight beautification in the emitted code and reduces verbosity at
the call site.

Also this technically fixes the indent in some cases (specifically when
serializing some repeated fields) so that the opening braces are emitted "at
the right level".
(Previous implementations sometimes emit an extra space character.)

Compare:
 {
  // code ...
}

vs.
{
  // code ...
}

PiperOrigin-RevId: 660236451