GitHub workflow to publish an artifactbundle for SwiftPM
# Motivation

We want to be able to consume protoc in the Swift ecosystem similar to how other ecosystems consume it. The goal is to allow seamless generation of Swift files from `proto` files via Swift PM build plugins. Currently, both `swift-protobuf` and `grpc-swift` offer such plugins; however, the usage remains limited due to the missing `protoc` dependency in the Swift PM ecosystem.

# Modification

This PR adds a GitHub workflow that is triggered when a release is created. The workflow fetches the released protoc artifacts, bundles them into an (artifactbundle)[https://github.com/swiftlang/swift-evolution/blob/main/proposals/0305-swiftpm-binary-target-improvements.md] and uploads them to the release. Once uploaded we can start to depend on them from `swift-protobuf` and offer the `protoc` binary to the Swift ecosystem.

# Result

Easy and seamless integration of protobuf based generation in the Swift ecosystem.