Add support for ignoring unknown enum values in text_format.Parse.
This is useful for parsing text protos that contain enum values that are not yet known to the binary. For example, this can be used to parse text protos that contain new enum values that have been added to a proto file but that have not yet been released in a binary.

PiperOrigin-RevId: 663322312