Put AnyMetadata logic into protoc.
This allows proto3's "Any" implementation to use the message's internal
accessors directly, instead of having an indirection to AnyMetadata. This also
plays well with hasbits because the setting and reading of hasbits can be
toggled via the internal accessors.

As a first iteration, I did not make radical changes to AnyMetadata to
hopefully make the review easier -- that is, I did not remove unused functions
in AnyMetadata. Since this change removes the need to depend on many of
AnyMetadata's internal functions, a logical next step would be to make the
AnyMetadata class into a pure wrapper class. Many functions can then be removed
in any.h, any.cc, any_lite.cc next. any_lite.cc can probably be removed entirely.

PiperOrigin-RevId: 663834705