Remove outdated conformance/fuzz tests.
These are very old unit-tests that appear to have been written to provide some conformance and fuzzing capabilities before we had more comprehensive solutions.  Today, they should be already covered by our conformance tests and fuzz tests and create unnecessary noise.

The goal of this change is to remove the checked-in binary golden data that we no longer know how to regenerate.  There is a lot of nearby code that could likely also be cleaned up in a follow-up

PiperOrigin-RevId: 667604310