Add map test cases to no_field_presence_test.
Maps behave in a very interesting way if they are set to be implicit-presence
fields.

First of all they present the possibility that the key or value (or both) can
be the default zero value. When that happens, the gencode behaviour and the
reflection behaviour is actually different.

For example, gencode would not count zero entries in the map. However,
reflection APIs such as HasField will return true for zero map entry fields. In
essence, for maps, reflection behaves like explicit presence, but gencode
behaves like implicit presence.

(Note that this is not exactly *intended*, but it's just hard to change...)

I'm adding unit tests in no_field_presence_test to cover this quirk.

PiperOrigin-RevId: 671126144