Rust: create blanket implementation of `ProxiedInMapValue` for upb-backed messages
This turned out to be much easier for upb than for C++. This CL pretty much
just does a cut-and-paste of the `ProxiedInMapValue` implementation from the
upb code generator to the runtime. This should help a bit with code size since
it removes the need to generate six `ProxiedInMapValue` implementations per
message.

PiperOrigin-RevId: 671438289