Rust: make generated views implement `std::default::Default`
The immediate motivation for this is that it will facilitate writing a blanket
implementation of `ProxiedInMapValue` for C++-backed messages. The default
instance gives us access to the message vtable in cases where we don't already
have a message to work with.

However, it also seems generally useful just to have an implementation of
`Default`, so I implemented it for both C++ and upb-backed message views.

PiperOrigin-RevId: 677808048