Add test cases for zero map entries for text format in Python.
This is just part of an effort at locking down (and understand well) how map
entries are serialized into textproto when they have zero fields.

Right now, the behaviour is that zero MapEntry fields, even if explicitly set,
are not serialized. This contrasts with C++ map reflection behaviour, because
C++ MapEntry messages are always generated with hasbits (even if they are
declared in a proto3 file and have implicit presence according to their
descriptors).

PiperOrigin-RevId: 678448520