[ObjC] Disabling visionOS on CI.
Github dropped support for the current image:
https://github.com/actions/runner-images/issues/10559

There is a new image that has visionOS, but it is listed as
"beta", so it might make sense to wait and revisit when
generally updating the macOS images/tools.

PiperOrigin-RevId: 679264196

[ObjC] Disabling visionOS on CI.

Github dropped support for the current image:
https://github.com/actions/runner-images/issues/10559

There is a new image that has visionOS, but it is listed as
"beta", so it might make sense to wait and revisit when
generally updating the macOS images/tools.

PiperOrigin-RevId: 679264196

[ObjC] Disabling visionOS on CI.

Github dropped support for the current image:
https://github.com/actions/runner-images/issues/10559

There is a new image that has visionOS, but it is listed as
"beta", so it might make sense to wait and revisit when
generally updating the macOS images/tools.

PiperOrigin-RevId: 679264196