Speed up CodedOutputStream by extracting rarely-executed string formatting code
If this code is in every method, it's making every method larger, putting
pressure on the instruction cache.

The string formatting is an exceptional case that shouldn't slow down the
regular case.

This should modestly speed up the code in here. And besides, it's just a little
nicer to have this formatting centralised.

PiperOrigin-RevId: 679336304