Add test for ListFields in implicit presence messages.
ListFields uses a slightly different codepath from HasField (both are
implemented in generated_message_reflection.cc), and it would be valuable to
add a test case for ListFields in addition to HasFields when reflection API are
tested.

ListFields is used, for example, in Python textproto serialization as a way to
iterate a message.

ListField behaves slightly differently for normal messages vs. MapEntry
messages. In a map, MapEntry essentially have "explicit presence" and empty
fields are considered present.

PiperOrigin-RevId: 682504379