Remove use of COPTS variable for C++ compiler flags
It's somewhat tedious to explicitly set this option on all of our C++ targets,
so I think it makes more sense to move it to our bazelrc files.

`-DHAVE_ZLIB` is the only option not related to warnings. Since we only have
two source files that rely on that, I just set that option only on the two
affected build targets.

PiperOrigin-RevId: 684569471