Rust: fix extra copy in map setters
When you call a map field setter, we currently make an unnecessary extra copy,
so this CL fixes that problem.

I followed the example of how we already handle this for repeated field
setters. This required adding a new move setter thunk for map fields with the
C++ kernel. Originally I tried instead to add an FFI endpoint that could swap
two `RawMap` pointers, but it turned out to be difficult to implement this in a
way that worked correctly when the two maps are not on the same arena.

PiperOrigin-RevId: 686272052