Bump abseil dependency version from 20230802.1 to 20240722.0.
See abseil release notes:
https://github.com/abseil/abseil-cpp/releases/tag/20240722.0

This is necesary pre-work to enable C++17 builds in preparation of baseline C++
upgrade from C++14 to C++17. C++14 support is scheduled to reach EOL on
2024-12-15:
https://github.com/google/oss-policies-info/blob/main/foundational-cxx-support-matrix.md

Interestingly enough, the older Abseil version 20230802.1 is actually OK for
building on macOS with C++17. However, we depend on GoogleTest, and they
transitively depend on Abseil version 20240116.2, which caused Bazel to
implicitly change the Abseil dependency to most recent specified version of
20240116.2:
https://github.com/google/googletest/blob/main/MODULE.bazel

Current builds with C++17 and Abseil version 20240116.2 on older macos versions
that didn't support `std::filesystem::path` would give the following error:

> error: 'path' is unavailable: introduced in macOS 10.15.

With this version upgrade, we pull in an Abseil patch that fixes this issue on
macOS:
https://github.com/abseil/abseil-cpp/commit/65a55c2ba891f6d2492477707f4a2e327a0b40dc
PiperOrigin-RevId: 686670268