Rust C++: get all map fields onto a common implementation of ProxiedInMapValue
This CL migrates messages, enums, and primitive types all onto the same blanket
implementation of the `ProxiedInMapValue` trait. This gets us to the point
where messages and enums no longer need to generate any significant amount of
extra code just in case they might be used as a map value.

There are a few big pieces to this:
 - I generalized the message-specific FFI endpoints in `rust/cpp_kernel/map.cc`
   to be able to additionally handle enums and primitive types as values. This
   mostly consisted of replacing `MessageLite*` parameters with a new `MapValue`
   tagged union.
 - On the Rust side, I added a new blanket implementation of
   `ProxiedInMapValue` in rust/cpp.rs. It relies on its value type to implement
   a new `CppMapTypeConversions` trait so that it can convert to and from the
   `MapValue` tagged union used for FFI.
 - In the Rust generated code, I deleted the generated `ProxiedInMapValue`
   implementations for messages and enums and replaced them with
   implementations of the `CppMapTypeConversions` trait.

PiperOrigin-RevId: 687355817