Rust: move proxy implementations of AssociatedMiniTable into the runtime
We are currently generating separate implementations of the
`AssociatedMiniTable` trait for owned messages, view proxies, and mutable
proxies. I don't think this is really necessary, though, because once we have
an implementation for one of these, we can derive it for the other two.

This CL deletes the proxy implementations from the generated code and adds a
blanket implementation to replace them. This is not super important, but it
seems nice to reduce our reliance on generated code a little bit.

PiperOrigin-RevId: 694680222