hpb: Improve ExtensionRegistry ergonomics
We introduce AddExtension, which'll take an ExtensionId. From there, our helper extracts the mini_table_ext() ptr and handles the upb_ExtensionRegistry behind the scenes.

The ctor for an ExtensionRegistry has been updated to just take an arena.

Just need to supply an extension, and let us handle the rest!

PiperOrigin-RevId: 696136849