Introduce hpb plumbing for repeated scalars
Before this change, hpb had no way of returning scalar repeated fields (that are extensions) -- they were incorrectly treated as pure scalars (int32 vs repeated<int32>).
To rectify this, we introduce RepeatedWrapper, which'll signal that we've detected a repeated scalar.

In a followup CL, we'll handle this case and properly return the extant RepeatedField, allowing access and iteration.
Just handling the non-mutational pathway for our first pass.

PiperOrigin-RevId: 696648468