Make fuzz test check round-trip correctness of upb encoder/decoder.
Prior to this CL, the fuzz tests only checked that the code does not crash, but it was not checking any correctness properties.  This CL adds correctness checking, verifying that we can round trip through the wire format without losing or corrupting data.

This highlighted a minor bug in the encoder where the depth limit check was off by one (too strict).  This CL makes the encoder more accepting when checking the depth limit.

PiperOrigin-RevId: 698527429