Backport windows fixes with INFINITY/NAN (#19389)
* Fixed `NAN`/`INFINITY` definitions to work on Windows 11 SDK.

Fixes: https://github.com/protocolbuffers/protobuf/issues/17308, https://github.com/protocolbuffers/protobuf/issues/17238

Confirmation that this addresses the issue: https://github.com/grpc/grpc/issues/37210#issuecomment-2226809722

PiperOrigin-RevId: 652541857

* Fix compilation in Clang C23 mode.

In C23, `false` is now of type `bool`, not `int`. This turned a couple places where `false` was returned instead of `NULL` into errors.

The workaround in upb for Windows's broken NAN macro causes an error in Clang under C23, `cannot compile this static initializer yet`, I believe because `0.0 / 0.0` is not valid in constant-evaluation. This seems like it's probably a legitimate result of C23 standardizing constexpr, although it _could_ be a clang bug. In any case, refine the workaround a bit, to avoid this problem.

I've also reverted the kUpb_FltInfinity/kUpb_Infinity back to their former definitions, as INFINITY wasn't broken by the windows header, only NAN.

PiperOrigin-RevId: 683152452

* Fix const INFINITY issue on MSVC.

We previously hit this for NAN, but it appears the latest version of MSVC used in windows github runners has the same issue for INFINITY.  The same strategy can be applied here.

This will be backported to fix broken release branches.

PiperOrigin-RevId: 699390594

* Regenerate stale files

* Update GitHub Actions to use macOS-13 ahead of macOS-12 deprecation.

See https://github.com/actions/runner-images/issues/10721.

PiperOrigin-RevId: 684546079

* Pin to python 3.12 in upb/macos tests.

Python 3.13 seems to have some issues with our dependencies and both numpy and setuptools fails to install.

PiperOrigin-RevId: 689955916

* Fix bad cherry-pick

---------

Co-authored-by: Joshua Haberman <haberman@google.com>
Co-authored-by: Protobuf Team Bot <protobuf-github-bot@google.com>